+++ e2e-tests/constants/routes/insurance/index.js
@@ -12,8 +12,6 @@ export const INSURANCE_ROUTES = {
     CHECK_IF_ELIGIBLE: `${INSURANCE_ROOT}${ELIGIBILITY}/check-if-eligible`,
     BUYER_COUNTRY: `${INSURANCE_ROOT}${ELIGIBILITY}/buyer-location`,
     CANNOT_APPLY: `${INSURANCE_ROOT}${ELIGIBILITY}/cannot-apply`,
-    APPLY_OFFLINE: `${INSURANCE_ROOT}${ELIGIBILITY}/apply-using-our-form`,
-    SPEAK_TO_UKEF_EFM: `${INSURANCE_ROOT}${ELIGIBILITY}/speak-to-UKEF-EFM`,
     EXPORTER_LOCATION: `${INSURANCE_ROOT}${ELIGIBILITY}/exporter-location`,
     UK_GOODS_OR_SERVICES: `${INSURANCE_ROOT}${ELIGIBILITY}/uk-goods-services`,
     INSURED_AMOUNT: `${INSURANCE_ROOT}${ELIGIBILITY}/insured-amount`,
@@ -26,6 +24,8 @@ export const INSURANCE_ROUTES = {
     ALREADY_HAVE_ACCOUNT: `${INSURANCE_ROOT}${ELIGIBILITY}/already-have-account`,
     NEED_TO_START_AGAIN: `${INSURANCE_ROOT}${ELIGIBILITY}/need-to-start-again`,
   },
+  SPEAK_TO_UKEF_EFM: `${INSURANCE_ROOT}/speak-to-UKEF-EFM`,
+  APPLY_OFFLINE: `${INSURANCE_ROOT}/apply-using-our-form`,
   ALL_SECTIONS: '/all-sections',
   PAGE_NOT_FOUND: `${INSURANCE_ROOT}/page-not-found`,
   POLICY_AND_EXPORTS,
