+++ e2e-tests/content-strings/pages/insurance/eligibility/index.js
@@ -8,58 +8,6 @@ export const MAX_COVER_AMOUNT = PRODUCT.MAX_COVER_AMOUNT_IN_GBP.toLocaleString('
   maximumFractionDigits: 0,
 });
 
-export const APPLY_OFFLINE = {
-  PAGE_TITLE: 'You need to apply using our form',
-  REASON: {
-    INTRO: 'This is because',
-    WANT_COVER_OVER_MAX_AMOUNT: `you want to be insured for more than ${MAX_COVER_AMOUNT} and we need to make extra checks.`,
-    OTHER_PARTIES_INVOLVED: 'there are other parties involved in your exports and we need to make extra checks.',
-    WILL_BE_PAID_BY_LETTER_OF_CREDIT: "you'll be paid by a letter of credit.",
-    NEED_PRE_CREDIT_PERIOD_COVER: 'you need pre-credit cover.',
-    NO_COMPANIES_HOUSE_NUMBER: 'you do not have a UK Companies House registration number',
-  },
-  ACTIONS: {
-    DOWNLOAD_FORM: {
-      LINK: {
-        TEXT: 'Download this form',
-        HREF: LINKS.EXTERNAL.NBI_FORM,
-      },
-      TEXT: "and email it to UKEF once you've filled it in.",
-    },
-    CONTACT: {
-      TEXT: "If you'd like to discuss your exports or need help applying, you can",
-      LINK: {
-        TEXT: 'talk to your nearest export finance manager.',
-        HREF: LINKS.EXTERNAL.EXPORT_FINANCE_MANAGERS,
-      },
-    },
-  },
-};
-
-export const SPEAK_TO_UKEF_EFM = {
-  PAGE_TITLE: 'You need to speak with a UKEF export finance manager',
-  REASON: {
-    INTRO: 'This is because',
-    WANT_COVER_OVER_MAX_PERIOD: `you want to be insured for longer than ${PRODUCT.MAX_COVER_PERIOD_YEARS} years.`,
-  },
-  ACTIONS: {
-    FIND_EFM: [
-      [
-        {
-          text: 'Find ',
-        },
-        {
-          text: 'your nearest export finance manager',
-          href: LINKS.EXTERNAL.EXPORT_FINANCE_MANAGERS,
-        },
-        {
-          text: ' to discuss this.',
-        },
-      ],
-    ],
-  },
-};
-
 export const CHECK_IF_ELIGIBLE = {
   PAGE_TITLE: 'Check you can apply for UKEF insurance for your export',
   BODY: 'This will take a couple of minutes. If your export is eligible, you can start the application immediately.',
