+++ e2e-tests/content-strings/pages/insurance/index.js
@@ -1,6 +1,71 @@
 import POLICY_AND_EXPORTS from './policy-and-exports';
 import * as ELIGIBILITY_PAGES from './eligibility';
 import * as EXPORTER_BUSINESS from './exporter-business';
+import { LINKS } from '../../links';
+import { PRODUCT } from '../../../constants';
+
+export const MAX_COVER_AMOUNT = PRODUCT.MAX_COVER_AMOUNT_IN_GBP.toLocaleString('en', {
+  style: 'currency',
+  currency: 'GBP',
+  minimumFractionDigits: 0,
+  maximumFractionDigits: 0,
+});
+
+const ALL_SECTIONS = {
+  PAGE_TITLE: 'Apply for UKEF export insurance',
+};
+
+const APPLY_OFFLINE = {
+  PAGE_TITLE: 'You need to apply using our form',
+  REASON: {
+    INTRO: 'This is because',
+    WANT_COVER_OVER_MAX_AMOUNT: `you want to be insured for more than ${MAX_COVER_AMOUNT} and we need to make extra checks.`,
+    OTHER_PARTIES_INVOLVED: 'there are other parties involved in your exports and we need to make extra checks.',
+    WILL_BE_PAID_BY_LETTER_OF_CREDIT: "you'll be paid by a letter of credit.",
+    NEED_PRE_CREDIT_PERIOD_COVER: 'you need pre-credit cover.',
+    NO_COMPANIES_HOUSE_NUMBER: 'you do not have a UK Companies House registration number',
+  },
+  ACTIONS: {
+    DOWNLOAD_FORM: {
+      LINK: {
+        TEXT: 'Download this form',
+        HREF: LINKS.EXTERNAL.NBI_FORM,
+      },
+      TEXT: "and email it to UKEF once you've filled it in.",
+    },
+    CONTACT: {
+      TEXT: "If you'd like to discuss your exports or need help applying, you can",
+      LINK: {
+        TEXT: 'talk to your nearest export finance manager.',
+        HREF: LINKS.EXTERNAL.EXPORT_FINANCE_MANAGERS,
+      },
+    },
+  },
+};
+
+const SPEAK_TO_UKEF_EFM = {
+  PAGE_TITLE: 'You need to speak with a UKEF export finance manager',
+  REASON: {
+    INTRO: 'This is because',
+    WANT_COVER_OVER_MAX_PERIOD: `you want to be insured for longer than ${PRODUCT.MAX_COVER_PERIOD_YEARS} years.`,
+  },
+  ACTIONS: {
+    FIND_EFM: [
+      [
+        {
+          text: 'Find ',
+        },
+        {
+          text: 'your nearest export finance manager',
+          href: LINKS.EXTERNAL.EXPORT_FINANCE_MANAGERS,
+        },
+        {
+          text: ' to discuss this.',
+        },
+      ],
+    ],
+  },
+};
 
 const START = {
   PAGE_TITLE: 'Apply for UKEF export insurance',
@@ -20,16 +85,14 @@ const START = {
   BODY_4: 'If you need it more urgently, you can tell us before you submit your application.',
 };
 
-const ALL_SECTIONS = {
-  PAGE_TITLE: 'Apply for UKEF export insurance',
-};
-
 const INSURANCE = {
-  START,
+  ALL_SECTIONS,
+  APPLY_OFFLINE,
   EXPORTER_BUSINESS,
   ELIGIBILITY: ELIGIBILITY_PAGES,
-  ALL_SECTIONS,
   POLICY_AND_EXPORTS,
+  SPEAK_TO_UKEF_EFM,
+  START,
 };
 
 export default INSURANCE;
