+++ e2e-tests/cypress/e2e/journeys/insurance/eligibility/pre-credit-period/pre-credit-period-answer-yes.spec.js
@@ -16,7 +16,7 @@ import {
   completeLetterOfCreditForm,
 } from '../../../../../support/insurance/eligibility/forms';
 
-const CONTENT_STRINGS = PAGES.INSURANCE.ELIGIBILITY.APPLY_OFFLINE;
+const CONTENT_STRINGS = PAGES.INSURANCE.APPLY_OFFLINE;
 
 context('Insurance - Eligibility - Pre-credit period page - I want to check if I can use online service to apply for UKEF Export Insurance Policy for my export transaction that is paid via letter of credit - submit `need pre-credit period cover`', () => {
   before(() => {
@@ -47,7 +47,7 @@ context('Insurance - Eligibility - Pre-credit period page - I want to check if I
   });
 
   it('redirects to exit page', () => {
-    cy.url().should('include', ROUTES.INSURANCE.ELIGIBILITY.APPLY_OFFLINE);
+    cy.url().should('include', ROUTES.INSURANCE.APPLY_OFFLINE);
   });
 
   it('renders a back link with correct url', () => {
