+++ src/ui/server/constants/routes/insurance/index.ts
@@ -13,8 +13,6 @@ export const INSURANCE_ROUTES = {
     CHECK_IF_ELIGIBLE: `${INSURANCE_ROOT}${ELIGIBILITY_ROOT}/check-if-eligible`,
     BUYER_COUNTRY: `${INSURANCE_ROOT}${ELIGIBILITY_ROOT}/buyer-location`,
     CANNOT_APPLY: `${INSURANCE_ROOT}${ELIGIBILITY_ROOT}/cannot-apply`,
-    APPLY_OFFLINE: `${INSURANCE_ROOT}${ELIGIBILITY_ROOT}/apply-using-our-form`,
-    SPEAK_TO_UKEF_EFM: `${INSURANCE_ROOT}${ELIGIBILITY_ROOT}/speak-to-UKEF-EFM`,
     EXPORTER_LOCATION: `${INSURANCE_ROOT}${ELIGIBILITY_ROOT}/exporter-location`,
     UK_GOODS_OR_SERVICES: `${INSURANCE_ROOT}${ELIGIBILITY_ROOT}/uk-goods-services`,
     INSURED_AMOUNT: `${INSURANCE_ROOT}${ELIGIBILITY_ROOT}/insured-amount`,
@@ -27,6 +25,8 @@ export const INSURANCE_ROUTES = {
     ALREADY_HAVE_ACCOUNT: `${INSURANCE_ROOT}${ELIGIBILITY_ROOT}/already-have-account`,
     NEED_TO_START_AGAIN: `${INSURANCE_ROOT}${ELIGIBILITY_ROOT}/need-to-start-again`,
   },
+  APPLY_OFFLINE: `${INSURANCE_ROOT}/apply-using-our-form`,
+  SPEAK_TO_UKEF_EFM: `${INSURANCE_ROOT}/speak-to-UKEF-EFM`,
   ALL_SECTIONS: '/all-sections',
   PAGE_NOT_FOUND: `${INSURANCE_ROOT}/page-not-found`,
   EXPORTER_BUSINESS,
