+++ src/ui/server/constants/templates/insurance/eligibility/index.ts
@@ -1,7 +1,5 @@
 export const ELIGIBILITY_TEMPLATES = {
   CHECK_IF_ELIGIBLE: 'insurance/eligibility/check-if-eligible.njk',
-  APPLY_OFFLINE: 'insurance/eligibility/apply-offline.njk',
-  SPEAK_TO_UKEF_EFM: 'insurance/eligibility/speak-to-ukef-efm.njk',
   UK_GOODS_OR_SERVICES: 'insurance/eligibility/uk-goods-or-services.njk',
   INSURED_AMOUNT: 'insurance/eligibility/insured-amount.njk',
   INSURED_PERIOD: 'insurance/eligibility/insured-period.njk',
