+++ src/ui/server/constants/templates/insurance/index.ts
@@ -5,8 +5,10 @@ import { BUSINESS_TEMPLATES } from './business';
 export const INSURANCE_TEMPLATES = {
   START: 'insurance/start.njk',
   ELIGIBILITY: ELIGIBILITY_TEMPLATES,
-  EXPORTER_BUSINESS: BUSINESS_TEMPLATES,
+  APPLY_OFFLINE: 'insurance/apply-offline.njk',
+  SPEAK_TO_UKEF_EFM: 'insurance/speak-to-ukef-efm.njk',
   ALL_SECTIONS: 'insurance/all-sections.njk',
   POLICY_AND_EXPORTS: POLICY_AND_EXPORTS_TEMPLATES,
+  EXPORTER_BUSINESS: BUSINESS_TEMPLATES,
   PAGE_NOT_FOUND: 'insurance/page-not-found.njk',
 };
