+++ src/ui/server/controllers/insurance/eligibility/insured-amount/index.test.ts
@@ -57,16 +57,16 @@ describe('controllers/insurance/eligibility/insured-amount', () => {
         };
       });
 
-      it(`should redirect to ${ROUTES.INSURANCE.ELIGIBILITY.APPLY_OFFLINE}`, async () => {
+      it(`should redirect to ${ROUTES.INSURANCE.APPLY_OFFLINE}`, async () => {
         await post(req, res);
 
-        expect(res.redirect).toHaveBeenCalledWith(ROUTES.INSURANCE.ELIGIBILITY.APPLY_OFFLINE);
+        expect(res.redirect).toHaveBeenCalledWith(ROUTES.INSURANCE.APPLY_OFFLINE);
       });
 
       it('should add exitReason to req.flash', async () => {
         await post(req, res);
 
-        const expectedReason = PAGES.INSURANCE.ELIGIBILITY.APPLY_OFFLINE.REASON.WANT_COVER_OVER_MAX_AMOUNT;
+        const expectedReason = PAGES.INSURANCE.APPLY_OFFLINE.REASON.WANT_COVER_OVER_MAX_AMOUNT;
         expect(req.flash).toHaveBeenCalledWith('exitReason', expectedReason);
       });
     });
