+++ src/ui/server/controllers/insurance/eligibility/other-parties/index.test.ts
@@ -57,16 +57,16 @@ describe('controllers/insurance/eligibility/other-parties', () => {
         };
       });
 
-      it(`should redirect to ${ROUTES.INSURANCE.ELIGIBILITY.APPLY_OFFLINE}`, async () => {
+      it(`should redirect to ${ROUTES.INSURANCE.APPLY_OFFLINE}`, async () => {
         await post(req, res);
 
-        expect(res.redirect).toHaveBeenCalledWith(ROUTES.INSURANCE.ELIGIBILITY.APPLY_OFFLINE);
+        expect(res.redirect).toHaveBeenCalledWith(ROUTES.INSURANCE.APPLY_OFFLINE);
       });
 
       it('should add exitReason to req.flash', async () => {
         await post(req, res);
 
-        const expectedReason = PAGES.INSURANCE.ELIGIBILITY.APPLY_OFFLINE.REASON.OTHER_PARTIES_INVOLVED;
+        const expectedReason = PAGES.INSURANCE.APPLY_OFFLINE.REASON.OTHER_PARTIES_INVOLVED;
         expect(req.flash).toHaveBeenCalledWith('exitReason', expectedReason);
       });
     });
