+++ src/ui/server/controllers/insurance/eligibility/other-parties/index.ts
@@ -35,12 +35,12 @@ const post = (req: Request, res: Response) => {
 
   if (answer === 'true') {
     const { INSURANCE } = PAGES;
-    const { APPLY_OFFLINE } = INSURANCE.ELIGIBILITY;
+    const { APPLY_OFFLINE } = INSURANCE;
     const { REASON } = APPLY_OFFLINE;
 
     req.flash('exitReason', REASON.OTHER_PARTIES_INVOLVED);
 
-    return res.redirect(ROUTES.INSURANCE.ELIGIBILITY.APPLY_OFFLINE);
+    return res.redirect(ROUTES.INSURANCE.APPLY_OFFLINE);
   }
 
   req.session.submittedData = {
