+++ src/ui/server/controllers/insurance/speak-to-ukef-efm/index.ts
@@ -1,0 +1,18 @@
+import { PAGES } from '../../../content-strings';
+import { TEMPLATES } from '../../../constants';
+import { Request, Response } from '../../../../types';
+import corePageVariables from '../../../helpers/page-variables/core/insurance';
+
+const get = (req: Request, res: Response) => {
+  const EXIT_REASON = req.flash('exitReason');
+
+  return res.render(TEMPLATES.INSURANCE.SPEAK_TO_UKEF_EFM, {
+    ...corePageVariables({
+      PAGE_CONTENT_STRINGS: PAGES.INSURANCE.SPEAK_TO_UKEF_EFM,
+      BACK_LINK: req.headers.referer,
+    }),
+    EXIT_REASON,
+  });
+};
+
+export default get;
