+++ src/ui/server/middleware/required-data-provided/insurance/eligibility/index.ts
@@ -2,10 +2,10 @@ import { FIELD_IDS, ROUTES } from '../../../../constants';
 import { Request, RequiredDataStateInsuranceEligibility, Response } from '../../../../../types';
 import { getRoutesAsArray, routeIsKnown, hasRequiredData } from '../../helpers';
 
+const { APPLY_OFFLINE, SPEAK_TO_UKEF_EFM } = ROUTES.INSURANCE;
+
 const {
   CANNOT_APPLY,
-  APPLY_OFFLINE,
-  SPEAK_TO_UKEF_EFM,
   CHECK_IF_ELIGIBLE,
   NEED_TO_START_AGAIN,
   ALREADY_HAVE_ACCOUNT,
