+++ src/ui/server/routes/insurance/index.ts
@@ -1,6 +1,8 @@
 import express from 'express';
 import { INSURANCE_ROUTES, INSURANCE_ROOT } from '../../constants/routes/insurance';
 import { get as startGet, post as startPost } from '../../controllers/insurance/start';
+import applyOfflineGet from '../../controllers/insurance/apply-offline';
+import speakToUkefEfmGet from '../../controllers/insurance/speak-to-ukef-efm';
 import { get as allSectionsGet } from '../../controllers/insurance/all-sections';
 import { get as typeOfPolicyGet, post as typeOfPolicyPost } from '../../controllers/insurance/policy-and-export/type-of-policy';
 import { get as singleContractPolicyGet } from '../../controllers/insurance/policy-and-export/single-contract-policy';
@@ -16,6 +18,9 @@ const insuranceRouter = express.Router();
 
 insuranceRouter.get(INSURANCE_ROUTES.START, startGet);
 insuranceRouter.post(INSURANCE_ROUTES.START, startPost);
+
+insuranceRouter.get(INSURANCE_ROUTES.APPLY_OFFLINE, applyOfflineGet);
+insuranceRouter.get(INSURANCE_ROUTES.SPEAK_TO_UKEF_EFM, speakToUkefEfmGet);
 
 insuranceRouter.get(`${INSURANCE_ROOT}/:referenceNumber${INSURANCE_ROUTES.ALL_SECTIONS}`, allSectionsGet);
